feat: add hardware-accelerated video wallpaper support
- Add GStreamer-based video player with NVIDIA NVDEC, AMD/Intel VAAPI support
- Implement DMA-BUF zero-copy rendering via zwp_linux_dmabuf_v1 protocol
- Add frame queue for decoupled GStreamer → Wayland rendering
- Add automatic H.264/HEVC → VP9 conversion for AMD hardware decode
- Add viewport-based GPU scaling for proper Zoom/Fit/Stretch modes
- Add graceful shutdown handling (SIGTERM/SIGINT) to prevent compositor freezes
- Add rayon parallel pixel conversion for wl_shm fallback path
- Respect FilterMethod config for image scaling quality

+++ src/colored.rs
@@ -28,12 +28,7 @@ pub fn gradient(
     let mut colors = Vec::with_capacity(gradient.colors.len());
 
     for &[r, g, b] in &*gradient.colors {
-        colors.push(colorgrad::Color::from_linear_rgba(
-            f32::from(r),
-            f32::from(g),
-            f32::from(b),
-            1.0,
-        ));
+        colors.push(colorgrad::Color::from_linear_rgba(r, g, b, 1.0));
     }
 
     let grad = colorgrad::GradientBuilder::new()
@@ -79,7 +74,7 @@ pub fn gradient(
     for (x, y, pixel) in imgbuf.enumerate_pixels_mut() {
         let Color { r, g, b, .. } = grad.at(positioner(x, y) as f32);
 
-        *pixel = image::Rgb([r as f32, g as f32, b as f32]);
+        *pixel = image::Rgb([r, g, b]);
     }
 
     Ok(imgbuf)
